+++ src/external/redacted/wxBTRFLY.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.6.12;
 
 abstract contract wxBTRFLY {
-    function realIndex() public view virtual returns (uint);
+    function realIndex() public view virtual returns (uint256);
 }
